requirements.txt
python-magic
sqlalchemy
